requirements.txt
streamlit==1.36.0
langchain==0.2.14
langchain-google-genai==1.0.8
langchain-community==0.2.10
pypdf==4.3.1
docx2txt==0.8
beautifulsoup4==4.12.3
chromadb==0.5.5
langchain-chroma==0.1.2
python-dotenv==1.0.1
